legalize-js-interface: always ensure getTempRet0/setTempRet0 exist
These helper functions are used by the llvm backend to implement
EH and sj/lj.  At least as of: https://reviews.llvm.org/D53240

Previously we were only generating them on demand if an i64 returning
function was exported.  Now we generate them in all cases.  DCE should
be able to remove them in the case when they are ultimately not needed.